+++ aqc_research/model_sp_lhs/objective_lhs_sur_fast_mps_trotter.py
@@ -108,7 +108,7 @@ class SpSurrogateObjectiveFastMpsTrotter(obj_base.SpLHSObjectiveBase):
             raise ValueError("only a single bit flip is currently supported")
 
         # Hilbert-Schmidt products <state|V.H|target> of the all states:
-        self._hs = np.zeros(self._num_states, dtype=np.cfloat)
+        self._hs = np.zeros(self._num_states, dtype=np.complex128)
 
         # Index of the state of maximal projection (max. HS product):
         self._max_no = int(0)
